fix: corrige les calculs sur les nombres décimaux

+++ impact/vsme/forms.py
@@ -204,7 +204,7 @@ def create_simple_field_from_schema(field_schema, rapport_vsme):
             field_kwargs["max_value"] = field_schema.get("max")
             field = forms.IntegerField(**field_kwargs)
         case "nombre_decimal":
-            field = forms.FloatField(**field_kwargs)
+            field = forms.DecimalField(**field_kwargs)
         case "date":
             field = forms.DateField(
                 widget=forms.DateInput(format="%Y-%m-%d", attrs={"type": "date"}),
